+++ rpc/lib/client/args_test.go
@@ -1,4 +1,4 @@
-package rpcclient
+package client
 
 import (
 	"testing"
