+++ rpc/lib/client/decode.go
@@ -1,4 +1,4 @@
-package rpcclient
+package client
 
 import (
 	"encoding/json"
