+++ rpc/lib/client/http_json_client.go
@@ -1,4 +1,4 @@
-package rpcclient
+package client
 
 import (
 	"bytes"
