+++ rpc/lib/client/http_uri_client.go
@@ -1,4 +1,4 @@
-package rpcclient
+package client
 
 import (
 	"fmt"
