+++ rpc/lib/client/integration_test.go
@@ -3,7 +3,7 @@
 // The code in here is comprehensive as an integration
 // test and is long, hence is only run before releases.
 
-package rpcclient
+package client
 
 import (
 	"bytes"
