+++ rpc/lib/client/ws_client.go
@@ -1,4 +1,4 @@
-package rpcclient
+package client
 
 import (
 	"context"
