+++ rpc/lib/server/http_json_handler_test.go
@@ -1,4 +1,4 @@
-package rpcserver
+package server
 
 import (
 	"bytes"
