+++ rpc/lib/server/http_params.go
@@ -1,4 +1,4 @@
-package rpcserver
+package server
 
 import (
 	"encoding/hex"
