+++ rpc/lib/server/http_server_test.go
@@ -1,4 +1,4 @@
-package rpcserver
+package server
 
 import (
 	"crypto/tls"
