+++ rpc/lib/server/http_uri_handler.go
@@ -1,4 +1,4 @@
-package rpcserver
+package server
 
 import (
 	"encoding/hex"
