+++ rpc/lib/server/parse_test.go
@@ -1,4 +1,4 @@
-package rpcserver
+package server
 
 import (
 	"encoding/json"
