+++ rpc/lib/server/rpc_func.go
@@ -1,4 +1,4 @@
-package rpcserver
+package server
 
 import (
 	"fmt"
