+++ rpc/lib/server/ws_handler.go
@@ -1,4 +1,4 @@
-package rpcserver
+package server
 
 import (
 	"context"
