+++ rpc/lib/server/ws_handler_test.go
@@ -1,4 +1,4 @@
-package rpcserver
+package server
 
 import (
 	"net/http"
