+++ rpc/lib/types/types.go
@@ -1,4 +1,4 @@
-package rpctypes
+package types
 
 import (
 	"context"
